+++ tests/node-test.js
@@ -9,3 +9,15 @@ md5 = null
 JS_MD5_TEST = true;
 require('../src/md5.js');
 require('./test.js');
+
+delete require.cache[require.resolve('../src/md5.js')];
+delete require.cache[require.resolve('./test.js')];
+md5 = null;
+
+define = function(func) {
+  md5 = func();
+  require('./test.js');
+};
+define.amd = true;
+
+require('../src/md5.js');